Refactor code into optimus package
