+++ data/optimus.txt
@@ -1,1 +1,1 @@
-T | 1 | deadline return book /by 2023-12-01E | 0 | meeting | 2023-12-01 | 2023-12-02
+T | 0 | borrow bookD | 0 | return book | 2020-12-20
