+++ src/main/java/optimus/Storage.java
@@ -1,3 +1,5 @@
+package optimus;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
