+++ src/main/java/optimus/Ui.java
@@ -1,3 +1,5 @@
+package optimus;
+
 import java.util.ArrayList;
 
 // Let ChatGPT check and suggest comments and JavaDocs according to CS2103T style guide
